Add simple js window confirmation to delete actions (#44)
component toggleSelect on delete button is because the click listener
exists on the whole div, so this resets the state if ‘cancel’ is
clicked in the confirmation dialog

+++ app/assets/javascripts/components/main/directives/course-card/courseCardCtrl.js
@@ -47,10 +47,12 @@
     }
 
     ctrl.delete = function () {
-      console.log('Deleting course with id: ' + $scope.gpCourse.id);
-      $http.delete('/courses/'+ $scope.gpCourse.id).then(function(){
-        $scope.gpParentCtrl.removeCourse($scope.gpCourse.id);
-      });
+      if(window.confirm("Are you sure you want to delete '" + $scope.gpCourse.code + " - " + $scope.gpCourse.name + "'?")){
+        console.log('Deleting course with id: ' + $scope.gpCourse.id);
+        $http.delete('/courses/'+ $scope.gpCourse.id).then(function(){
+          $scope.gpParentCtrl.removeCourse($scope.gpCourse.id);
+        });
+      }
     }
 
     ctrl.save = function () {
